requirements.txt
amqp==5.0.9
pykka==3.0.2
certifi==2021.10.8
wheel==0.37.0
inotify==0.2.10
nose==1.3.7
paho-mqtt==1.5.1
pkg-resources==0.0.0
vine==5.0.0
watchdog==2.1.6
